move of config class and small tweaks
